+++ internal/elasticsearch/ml/calendar_job/testdata/TestAccResourceMLCalendarJob_applyCalendarNotFound/create/calendar_job.tf
@@ -1,4 +1,4 @@
-variable "calendar_id" {
+variable "missing_calendar_id" {
   type = string
 }
 
@@ -12,7 +12,7 @@ provider "elasticstack" {
 
 resource "elasticstack_elasticsearch_ml_anomaly_detection_job" "job" {
   job_id      = var.job_id
-  description = "ACC job for ml_calendar_job"
+  description = "ACC job for ml_calendar_job apply missing calendar"
 
   analysis_config = {
     bucket_span = "15m"
@@ -30,6 +30,6 @@ resource "elasticstack_elasticsearch_ml_anomaly_detection_job" "job" {
 }
 
 resource "elasticstack_elasticsearch_ml_calendar_job" "test" {
-  calendar_id = var.calendar_id
+  calendar_id = var.missing_calendar_id
   job_id      = elasticstack_elasticsearch_ml_anomaly_detection_job.job.job_id
 }