+++ internal/elasticsearch/ml/calendar_job/testdata/TestAccResourceMLCalendarJob_withJobGroup/create/calendar_job.tf
@@ -6,13 +6,19 @@ variable "job_id" {
   type = string
 }
 
+variable "group_name" {
+  type = string
+}
+
 provider "elasticstack" {
   elasticsearch {}
 }
 
 resource "elasticstack_elasticsearch_ml_anomaly_detection_job" "job" {
   job_id      = var.job_id
-  description = "ACC job for ml_calendar_job"
+  description = "ACC job for ml_calendar_job group (groups set in HCL)"
+
+  groups = [var.group_name]
 
   analysis_config = {
     bucket_span = "15m"
@@ -31,5 +37,7 @@ resource "elasticstack_elasticsearch_ml_anomaly_detection_job" "job" {
 
 resource "elasticstack_elasticsearch_ml_calendar_job" "test" {
   calendar_id = var.calendar_id
-  job_id      = elasticstack_elasticsearch_ml_anomaly_detection_job.job.job_id
+  job_id      = var.group_name
+
+  depends_on = [elasticstack_elasticsearch_ml_anomaly_detection_job.job]
 }